+++ app/Models/Visit.php
@@ -67,63 +67,6 @@ class Visit extends Model
     }
 
     /**
-     * The number of clicks from links created by the currently authenticated
-     * user.
-     */
-    public function authUserLinkVisits(): int
-    {
-        return self::whereRelation('url', 'user_id', auth()->id())
-            ->count();
-    }
-
-    /**
-     * The number of clicks from links created by all registered users.
-     */
-    public function userLinkVisits(): int
-    {
-        return self::whereRelation('url', 'user_type', UserType::User)
-            ->count();
-    }
-
-    /**
-     * The number of clicks from links created by all guest users.
-     */
-    public function guestLinkVisits(): int
-    {
-        return self::whereRelation('url', 'user_type', UserType::Guest)
-            ->count();
-    }
-
-    /**
-     *  Total users who clicked on a link.
-     */
-    public function userVisits(): int
-    {
-        return self::where('user_type', UserType::User)->count();
-    }
-
-    /**
-     * Total guest users who clicked on a link.
-     *
-     * @param bool $unique Whether to count unique guest users or all guest visits.
-     * @return int
-     */
-    public function guestVisits(bool $unique = false)
-    {
-        return self::isGuest()
-            ->when($unique, fn($query) => $query->distinct('user_uid'))
-            ->count();
-    }
-
-    /**
-     * Total unique guest users who clicked on a link.
-     */
-    public function uniqueGuestVisits(): int
-    {
-        return $this->guestVisits(true);
-    }
-
-    /**
      * Get the top referrers based on visit count.
      *
      * @param \App\Models\User|\App\Models\Url|null $object Object to filter items.
