+++ app/Services/VisitService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services;
 
+use App\Enums\UserType;
 use App\Helpers\Helper;
 use App\Models\Url;
 use App\Models\Visit;
@@ -72,5 +73,62 @@ class VisitService
         $referer = Uri::of($value);
 
         return $referer->scheme().'://'.$referer->host();
+    }
+
+    /**
+     * The number of clicks from links created by the currently authenticated
+     * user.
+     */
+    public function authUserLinkVisits(): int
+    {
+        return Visit::whereRelation('url', 'user_id', auth()->id())
+            ->count();
+    }
+
+    /**
+     * The number of clicks from links created by all registered users.
+     */
+    public function userLinkVisits(): int
+    {
+        return Visit::whereRelation('url', 'user_type', UserType::User)
+            ->count();
+    }
+
+    /**
+     * The number of clicks from links created by all guest users.
+     */
+    public function guestLinkVisits(): int
+    {
+        return Visit::whereRelation('url', 'user_type', UserType::Guest)
+            ->count();
+    }
+
+    /**
+     *  Total users who clicked on a link.
+     */
+    public function userVisits(): int
+    {
+        return Visit::where('user_type', UserType::User)->count();
+    }
+
+    /**
+     * Total guest users who clicked on a link.
+     *
+     * @param bool $unique Whether to count unique guest users or all guest visits.
+     * @return int
+     */
+    public function guestVisits(bool $unique = false)
+    {
+        return Visit::isGuest()
+            ->when($unique, fn($query) => $query->distinct('user_uid'))
+            ->count();
+    }
+
+    /**
+     * Total unique guest users who clicked on a link.
+     */
+    public function uniqueGuestVisits(): int
+    {
+        return $this->guestVisits(true);
     }
 }
