+++ tests/Unit/Models/VisitTest.php
@@ -2,7 +2,6 @@
 
 namespace Tests\Unit\Models;
 
-use App\Enums\UserType;
 use App\Models\Url;
 use App\Models\User;
 use App\Models\Visit;
@@ -12,6 +11,8 @@ use Tests\TestCase;
 #[PHPUnit\Group('model')]
 class VisitTest extends TestCase
 {
+    private Visit $visit;
+
     protected function setUp(): void
     {
         parent::setUp();
@@ -33,96 +34,6 @@ class VisitTest extends TestCase
 
         $this->assertEquals(1, $visit->url->count());
         $this->assertInstanceOf(Url::class, $visit->url);
-    }
-
-    #[PHPUnit\Test]
-    public function authUserLinkVisits(): void
-    {
-        $user = $this->basicUser();
-        $nCurrentUser = 8;
-        $nUser = 6;
-
-        Visit::factory()->count($nCurrentUser)
-            ->for(Url::factory()->state(['user_id' => $user->id]))
-            ->create();
-        Visit::factory()->count($nUser)
-            ->for(Url::factory())
-            ->create();
-
-        $this->actingAs($user);
-        $this->assertSame($nCurrentUser, $this->visit->authUserLinkVisits());
-        $this->assertSame($nCurrentUser + $nUser, $this->visit->userLinkVisits());
-    }
-
-    #[PHPUnit\Test]
-    public function userLinkVisits(): void
-    {
-        $nUser = 6;
-        $nGuest = 4;
-
-        Visit::factory()->count($nUser)
-            ->for(Url::factory())
-            ->create();
-
-        Visit::factory()->count($nGuest)
-            ->for(Url::factory()->guest())
-            ->create();
-
-        $this->assertSame($nUser, $this->visit->userLinkVisits());
-        $this->assertSame($nUser + $nGuest, $this->visit->count());
-    }
-
-    #[PHPUnit\Test]
-    public function guestLinkVisits(): void
-    {
-        $nUser = 6;
-        $nGuest = 4;
-
-        Visit::factory()->count($nUser)
-            ->for(Url::factory())
-            ->create();
-
-        Visit::factory()->count($nGuest)
-            ->for(Url::factory()->guest())
-            ->create();
-
-        $this->assertSame($nGuest, $this->visit->guestLinkVisits());
-        $this->assertSame($nUser + $nGuest, $this->visit->count());
-    }
-
-    #[PHPUnit\Test]
-    public function userVisits()
-    {
-        $this->visitCountData();
-
-        $this->assertEquals(1, $this->visit->userVisits());
-    }
-
-    #[PHPUnit\Test]
-    public function guestVisits()
-    {
-        $this->visitCountData();
-
-        $this->assertEquals(5, $this->visit->guestVisits());
-    }
-
-    #[PHPUnit\Test]
-    public function uniqueGuestVisits()
-    {
-        $this->visitCountData();
-
-        $this->assertEquals(3, $this->visit->uniqueGuestVisits());
-    }
-
-    private function visitCountData()
-    {
-        Visit::factory()->create(); // user1
-        Visit::factory()->guest()->create(); // guest1
-        Visit::factory()->guest()->count(2)->create(['user_uid' => 'foo']); // guest2
-        Visit::factory()->count(2)->create([ // bot
-            'user_type' => UserType::Bot,
-            'user_uid' => 'bar',
-        ]);
     }
 
     public function test_get_top_referrers()
